webadmin/requirements.txt
asgiref==3.9.1
Django==4.2.23
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-js-asset==3.1.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==20.1.0
mysqlclient==2.2.7
PyJWT==2.9.0
PyMySQL==1.1.1
setuptools==80.9.0
sqlparse==0.5.3
